pyproject.toml
[tool.poetry]
name = "digest-emails"
version = "0.1.0"
description = ""
authors = ["Peter Stuart <peter.stuart@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.8"
openai = "0.28.1"
python-dotenv = "^1.0.0"
tiktoken = "^0.5.1"
pyqt6 = "^6.6.0"
pyinstaller = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
